Feature/cheyenne gnu

+++ libs/build_nceplibs.sh
@@ -22,6 +22,7 @@ openmp=${4:-${s_openmp:-"OFF"}}
 compiler=$(echo $HPC_COMPILER | sed 's/\//-/g')
 mpi=$(echo $HPC_MPI | sed 's/\//-/g')
 python=$(echo $HPC_PYTHON | sed 's/\//-/g')
+[[ -z $mpi ]] && modpath=compiler || modpath=mpi
 
 if $MODULES; then
   set +x
@@ -30,7 +31,7 @@ if $MODULES; then
 
   case $name in
     # The following require MPI
-    nemsiogfs | ncio | nceppost | upp)
+    nemsiogfs | ncio )
       module load hpc-$HPC_MPI
       using_mpi=YES
       ;;
@@ -83,29 +84,37 @@ if $MODULES; then
   # Load dependencies
   case $name in
     wrf_io)
-      module load netcdf
+      module restore hpc-$modpath-netcdf
+      module is-loaded netcdf || module load netcdf
       ;;
     wgrib2)
+      module restore hpc-$modpath-netcdf
+      module is-loaded netcdf || module load netcdf
       module try-load jpeg
       module try-load jasper
-      module try-load zlib
-      module try-load png
-      module load netcdf
+      module is-loaded zlib || module try-load zlib
+      module try-load libpng
       module load sp
       module load ip2
+      ;;
+    crtm)
+      module load hpc-$HPC_MPI
+      module restore hpc-$modpath-netcdf
+      module is-loaded netcdf || module load netcdf
       ;;
     ip2)
       module load sp
       ;;
     g2)
       module try-load jpeg
-      module try-load png
+      module try-load libpng
       module try-load jasper
       ;;
     g2c)
+
       module try-load jpeg
-      module try-load zlib
-      module try-load png
+      module is-loaded zlib || module try-load zlib
+      module try-load libpng
       module try-load jasper
       ;;
     nemsio)
@@ -119,35 +128,18 @@ if $MODULES; then
     w3emc)
       module load bacio
       if [[ "$using_mpi" =~ [yYtT] ]]; then
-          module load netcdf
+          module restore hpc-$modpath-netcdf
+          module is-loaded netcdf || module load netcdf
+          module load bacio
           module load sigio
           module load nemsio
       fi
-      ;;
-    nceppost | upp)
-      module try-load png
-      module try-load jasper
-      module load netcdf
-      module load bacio
-      module load w3nco
-      module load g2
-      module load g2tmpl
-      module load ip
-      module load sp
-      module load w3emc
-      module load crtm
-      # post executable requires the following,
-      # but we are not building post executable
-      # module load sigio
-      # module load sfcio
-      # module load gfsio
-      # module load nemsio
       ;;
     grib_util)
       module try-load jpeg
       module try-load jasper
       module try-load zlib
-      module try-load png
+      module try-load libpng
       module load bacio
       module load w3nco
       module load g2
@@ -178,8 +170,14 @@ if $MODULES; then
     prefix="${PREFIX:-"/opt/modules"}/$compiler/$name/$install_as"
   fi
   if [[ -d $prefix ]]; then
-    [[ $OVERWRITE =~ [yYtT] ]] && ( echo "WARNING: $prefix EXISTS: OVERWRITING!";$SUDO rm -rf $prefix; $SUDO mkdir $prefix ) \
-                               || ( echo "WARNING: $prefix EXISTS, SKIPPING"; exit 1 )
+      if [[ $OVERWRITE =~ [yYtT] ]]; then
+          echo "WARNING: $prefix EXISTS: OVERWRITING!"
+          $SUDO rm -rf $prefix
+          $SUDO mkdir $prefix
+      else
+          echo "WARNING: $prefix EXISTS, SKIPPING"
+          exit 0
+      fi
   fi
 
 else
@@ -188,7 +186,7 @@ else
   eval prefix="\${${nameUpper}_ROOT:-'/usr/local'}"
   case $name in
     # The following require MPI
-    nemsio | nemsiogfs | ncio | nceppost | upp)
+    nemsio | nemsiogfs | ncio )
       using_mpi=YES
       ;;
     # The following can use MPI (if available)
@@ -223,12 +221,8 @@ export FCFLAGS="$FFLAGS"
 URL="https://github.com/noaa-emc/nceplibs-$name"
 extraCMakeFlags=""
 case $name in
-  nceppost | upp)
-    URL="https://github.com/noaa-emc/upp"
-    extraCMakeFlags="-DBUILD_POSTEXEC=OFF"
-    ;;
   crtm)
-    URL="https://github.com/noaa-emc/emc_crtm"
+    URL="https://github.com/NOAA-EMC/crtm.git"
     ;;
   wgrib2)
     extraCMakeFlags="${STACK_wgrib2_cmake_opts:-}"
@@ -240,7 +234,7 @@ case $name in
     if [[ $MAKE_CHECK =~ [yYtT] ]]; then
         extraCMakeFlags+="-DBUILD_TESTS=ON"
     else
-        extraCMakeFlags+="-DBUILD_TETS=OFF"
+        extraCMakeFlags+="-DBUILD_TESTS=OFF"
     fi
     ;;
   nemsio)
@@ -256,8 +250,12 @@ cd ${HPC_STACK_ROOT}/${PKGDIR:-"pkg"}
 
 software=$name-$version
 if [[ ! -d $software ]]; then
+  export GIT_LFS_SKIP_SMUDGE=1
   git clone $URL $software
   cd $software
+  if [[ "$name" == "crtm" ]]; then
+    version=release/REL-${install_as}_emc
+  fi
   git checkout $version
   git submodule update --init --recursive
 fi
@@ -267,19 +265,20 @@ cd ${HPC_STACK_ROOT}/${PKGDIR:-"pkg"}
 # Download CRTM fix files
 if [[ "$name" == "crtm" ]]; then
   if [[ ${STACK_crtm_install_fix:-} =~ [yYtT] ]]; then
-    if [[ ! -d crtm_fix-$version ]]; then
+    if [[ ! -d crtm_fix-${install_as} ]]; then
       crtm_tarball=fix_REL-${install_as}_emc.tgz
-      rm -f $crtm_tarball
-      $WGET ftp://ftp.ucar.edu/pub/cpaess/bjohns/$crtm_tarball
+     [[ -f $crtm_tarball ]] || ( $WGET ftp://ftp.ucar.edu/pub/cpaess/bjohns/$crtm_tarball )
       tar xzf $crtm_tarball
-      mv fix crtm_fix-$version
-      rm -f $crtm_tarball
+      mv fix crtm_fix-${install_as}
+#      rm -f $crtm_tarball
     fi
-    if [[ ! -f link_crtm_coeffs.sh ]]; then
-      $WGET https://raw.githubusercontent.com/NOAA-EMC/GSI/master/ush/link_crtm_coeffs.sh
-      sed -i'.backup' -e 's/LINK="ln -sf"/LINK="cp"/g' link_crtm_coeffs.sh
-      chmod +x link_crtm_coeffs.sh
-      rm -f link_crtm_coeffs.sh.backup
+    if [[ "${install_as}" == "2.3.0" ]]; then
+     if [[ ! -f link_crtm_coeffs.sh ]]; then
+       $WGET https://raw.githubusercontent.com/NOAA-EMC/GSI/master/ush/link_crtm_coeffs.sh
+       sed -i'.backup' -e 's/LINK="ln -sf"/LINK="cp"/g' link_crtm_coeffs.sh
+       chmod +x link_crtm_coeffs.sh
+       rm -f link_crtm_coeffs.sh.backup
+     fi
     fi
   fi
 fi
@@ -305,8 +304,24 @@ cd ${HPC_STACK_ROOT}/${PKGDIR:-"pkg"}
 # Install CRTM fix files
 if [[ "$name" == "crtm" ]]; then
   if [[ ${STACK_crtm_install_fix:-} =~ [yYtT] ]]; then
-    if [[ -d crtm_fix-$version ]]; then
-      ./link_crtm_coeffs.sh ./crtm_fix-$version $prefix/fix
+    if [[ -d crtm_fix-${install_as} ]]; then
+     if [[ "${install_as}" == "2.3.0" ]]; then
+       ./link_crtm_coeffs.sh ./crtm_fix-${install_as} $prefix/fix
+     else
+       mkdir -p $prefix/fix
+       cp ./crtm_fix-${install_as}/ACCoeff/netcdf/* $prefix/fix
+       cp ./crtm_fix-${install_as}/AerosolCoeff/Big_Endian/* $prefix/fix
+       cp ./crtm_fix-${install_as}/AerosolCoeff/netCDF/* $prefix/fix
+       cp ./crtm_fix-${install_as}/CloudCoeff/Big_Endian/* $prefix/fix
+       cp ./crtm_fix-${install_as}/CloudCoeff/netCDF/* $prefix/fix
+       cp ./crtm_fix-${install_as}/EmisCoeff/*/Big_Endian/* $prefix/fix
+       cp ./crtm_fix-${install_as}/EmisCoeff/*/*/Big_Endian/* $prefix/fix
+       cp ./crtm_fix-${install_as}/SpcCoeff/Big_Endian/* $prefix/fix
+       cp ./crtm_fix-${install_as}/SpcCoeff/netcdf/* $prefix/fix
+       cp ./crtm_fix-${install_as}/TauCoeff/ODPS/Big_Endian/* $prefix/fix
+       mv $prefix/fix/amsua_metop-c.SpcCoeff.bin $prefix/fix/amsua_metop-c.SpcCoeff.noACC.bin
+       cp ./crtm_fix-${install_as}/SpcCoeff/Little_Endian/amsua_metop-c_v2.SpcCoeff.bin $prefix/fix/amsua_metop-c.SpcCoeff.bin
+     fi
     fi
   fi
 fi
